display: carve encode/pool types.rs (pure move)
